+++ source/C_API.cpp
@@ -92,6 +92,6 @@ extern "C" void TA_string_delete(void *const string_ptr) noexcept {
   delete static_cast<char *>(string_ptr);
 }
 
-extern "C" int64_t TA_get_storage_size(const int32_t datatype) noexcept {
+extern "C" int32_t TA_get_storage_size(const int32_t datatype) noexcept {
   return tagarray::utils::get_storage_size(datatype);
 }
